Finalised Rough Figure C plotting
Added Electron Temperature tests where Global Fitting is run for a range of base temps and the chi squared for the conductance and occupation data is reported.
Also updated the periodogram so that the integrated value is independent of scan duration and ADC speed.
Added choice to global fit for different gammas so initial coefficient guesses will vary based on gamma.
Updated the notch filtering frequencies and Q factors.